code cleanup

+++ src/particle_filter.hpp
@@ -3,19 +3,17 @@
 
 #include <vector>
 #include <cmath>
+
 #include "particle.hpp"
 #include "maplevel99plus.hpp"
 #include "utilities.hpp"
-
-
-
 
  void Particle_filter(std::vector<Particle>& particlesList, int U, double Z, double avg, double standardDeviation, int currentPosition, double arr[]) {
 	std::vector<Particle> newS;
 	MapLevel99Plus map;
 	map.buildForVector(particlesList);
 	double weightSum = 0;
-	for (int i = 0; i < particlesList.size(); i++) {
+	for (unsigned long i = 0; i < particlesList.size(); i++) {
 		Particle randomParticle = Particle(&utilities::getRandomParticle(map));
 		if (randomParticle.getPosition() + U >= 1000)
 			randomParticle.setPosition((rand() % (1000- currentPosition))+ currentPosition);
@@ -26,14 +24,11 @@
 		weightSum += weightParicle;
 		newS.push_back(randomParticle);
 	};
-	for (int i = 0; i < particlesList.size(); i++) {
+	for (unsigned long i = 0; i < particlesList.size(); i++) {
 		particlesList[i] = newS[i];
 	}
-	for (int i = 0; i < particlesList.size(); i++)
+	for (unsigned long i = 0; i < particlesList.size(); i++)
 		particlesList[i] = utilities::normalize(particlesList[i], weightSum);
-
-
-
 }
 
 #endif // PARTICLE_FILTER_HPP_INCLUDED
